+++ src/lib/src/index.ts
@@ -1,5 +1,6 @@
 export * from './config';
 export * from './dispatch-events';
+export * from './dom-selectors';
 export * from './element-focus';
 export * from './event-objects';
 export * from './host';
@@ -13,3 +14,4 @@ export * from './spectator';
 export * from './type-in-element';
 export * from './globals';
 export * from './mock-component';
+export * from './mock-directive';